wip: Add new window + routing for log window
