+++ src/pages/Logs.tsx
@@ -1,0 +1,5 @@
+const Logs = () => {
+  return <div className="log-window">Logs</div>;
+};
+
+export default Logs;
